LayoutTests/fast/dom/Window/a-rel-noopener-expected.txt

Test that window.opener is null when a new window is opened from an anchor element with rel='noopener'.

PASS: window.opener is null
